backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2

# AI Bias Detection
aif360==0.6.1
fairlearn==0.10.0
shap==0.46.0

# ML
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
scipy==1.14.0

# Claude AI
anthropic==0.34.2

# Database
supabase==2.7.4

# PDF Export
reportlab==4.2.2

# Utils
python-dotenv==1.0.1
httpx==0.27.2

google-generativeai==0.7.2